F #26: Starting with the integration

+++ Obento/Obento/Models/Ingredient.swift
@@ -13,7 +13,7 @@ struct Ingredient {
     var category: String
     var unitaryPrice: Double
     var unit: String
-    var kcal: Int
+    var kcal: Double
     var iconPath: String
     var quantity: Int? // Only for shopping list
 }
